requirements.txt
pandas
xlrd
openpyxl
pymupdf
numpy
pytesseract
pillow
opencv-python